Modified controls, improved mecanum functionality
I took a look at the teleop code and changed the controls on the gamepad in several ways.  I disabled the code on the bumpers that controlled rotation, and tweaked the strafing code to map to the d-pad. Regular drive was switched from the left to the right joystick with proper rotation controls, and I also added functionality to let the robot strafe diagonally by pushing the joystick forward or back while using the d-pad for left/right strafing. I also fixed the mecanum drive train code to make the speeds on both the left and the right sides consistent, because before, when moving left, the left wheels would move over 2x as fast as the right, and vice versa.